# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TVR Dubbers streaming API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: episodes
    description: Episode management
  - name: comments
    description: Comment system
  - name: reactions
    description: Reaction system
  - name: settings
    description: Site settings
  - name: trailer
    description: Trailer management
  - name: voice-artists
    description: Voice artist management
  - name: admin
    description: Admin authentication and management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # Episodes
  /episodes:
    get:
      operationId: listEpisodes
      tags: [episodes]
      summary: List all episodes
      parameters:
        - in: query
          name: genre
          schema:
            type: string
          required: false
        - in: query
          name: special
          schema:
            type: string
            enum: ["true", "false"]
          required: false
      responses:
        "200":
          description: List of episodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Episode"
    post:
      operationId: createEpisode
      tags: [episodes]
      summary: Create new episode (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpisodeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
        "401":
          description: Unauthorized

  /episodes/{id}:
    get:
      operationId: getEpisode
      tags: [episodes]
      summary: Get single episode
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Episode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
        "404":
          description: Not found
    patch:
      operationId: updateEpisode
      tags: [episodes]
      summary: Update episode (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpisodeUpdate"
      responses:
        "200":
          description: Updated episode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteEpisode
      tags: [episodes]
      summary: Delete episode (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /episodes/{id}/view:
    post:
      operationId: incrementEpisodeView
      tags: [episodes]
      summary: Increment view count
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated view count
          content:
            application/json:
              schema:
                type: object
                properties:
                  viewCount:
                    type: integer
                required: [viewCount]

  /episodes/stats/summary:
    get:
      operationId: getEpisodeStats
      tags: [episodes]
      summary: Get episode summary stats (total count, genre breakdown, recent)
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeStats"

  # Comments
  /episodes/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for episode
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add comment to episode
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "429":
          description: Too many requests

  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  /comments/all:
    get:
      operationId: listAllComments
      tags: [comments]
      summary: List all comments grouped by episode (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

  # Reactions
  /episodes/{id}/reactions:
    get:
      operationId: getReactions
      tags: [reactions]
      summary: Get reaction counts for episode
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionCounts"
    post:
      operationId: addReaction
      tags: [reactions]
      summary: Add or change reaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Updated reaction counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionCounts"

  # Trailer
  /trailer:
    get:
      operationId: getTrailer
      tags: [trailer]
      summary: Get current trailer
      responses:
        "200":
          description: Trailer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trailer"
    put:
      operationId: updateTrailer
      tags: [trailer]
      summary: Update trailer (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrailerInput"
      responses:
        "200":
          description: Updated trailer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trailer"
        "401":
          description: Unauthorized

  # Settings
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
        "401":
          description: Unauthorized

  # Voice Artists
  /voice-artists:
    get:
      operationId: listVoiceArtists
      tags: [voice-artists]
      summary: List voice artists
      responses:
        "200":
          description: Voice artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VoiceArtist"
    post:
      operationId: createVoiceArtist
      tags: [voice-artists]
      summary: Add voice artist (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceArtistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceArtist"
        "401":
          description: Unauthorized

  /voice-artists/{id}:
    delete:
      operationId: deleteVoiceArtist
      tags: [voice-artists]
      summary: Remove voice artist (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  # Admin Auth
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid password
        "429":
          description: Too many attempts

  /admin/change-password:
    post:
      operationId: changeAdminPassword
      tags: [admin]
      summary: Change admin password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
        "400":
          description: Current password incorrect
        "401":
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Episode:
      type: object
      required: [id, title, episodeNumber, season, createdAt, viewCount, isSpecial]
      properties:
        id:
          type: integer
        title:
          type: string
        episodeNumber:
          type: integer
        season:
          type: integer
        genre:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        primaryServerUrl:
          type: ["string", "null"]
        backupServerUrl:
          type: ["string", "null"]
        isSpecial:
          type: boolean
        createdAt:
          type: string
        viewCount:
          type: integer

    EpisodeInput:
      type: object
      required: [title, episodeNumber]
      properties:
        title:
          type: string
          minLength: 1
        episodeNumber:
          type: integer
        season:
          type: integer
        genre:
          type: string
        thumbnailUrl:
          type: string
        primaryServerUrl:
          type: string
        backupServerUrl:
          type: string
        isSpecial:
          type: boolean

    EpisodeUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        episodeNumber:
          type: integer
        season:
          type: integer
        genre:
          type: string
        thumbnailUrl:
          type: string
        primaryServerUrl:
          type: string
        backupServerUrl:
          type: string
        isSpecial:
          type: boolean

    EpisodeStats:
      type: object
      required: [totalEpisodes, totalSpecial, totalViews, genreCounts, recentEpisodes]
      properties:
        totalEpisodes:
          type: integer
        totalSpecial:
          type: integer
        totalViews:
          type: integer
        genreCounts:
          type: array
          items:
            type: object
            required: [genre, count]
            properties:
              genre:
                type: string
              count:
                type: integer
        recentEpisodes:
          type: array
          items:
            $ref: "#/components/schemas/Episode"

    Comment:
      type: object
      required: [id, episodeId, nickname, body, createdAt]
      properties:
        id:
          type: integer
        episodeId:
          type: integer
        nickname:
          type: string
        body:
          type: string
        createdAt:
          type: string

    CommentInput:
      type: object
      required: [body]
      properties:
        nickname:
          type: string
        body:
          type: string
          minLength: 1

    ReactionCounts:
      type: object
      required: [episodeId, counts, total]
      properties:
        episodeId:
          type: integer
        counts:
          type: object
          additionalProperties:
            type: integer
        total:
          type: integer

    ReactionInput:
      type: object
      required: [reactionType, visitorId]
      properties:
        reactionType:
          type: string
          enum: ["👍", "❤️", "👎", "🔥", "😥", "😹", "💀"]
        visitorId:
          type: string

    Trailer:
      type: object
      required: [id]
      properties:
        id:
          type: integer
        title:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        primaryServerUrl:
          type: ["string", "null"]
        backupServerUrl:
          type: ["string", "null"]

    TrailerInput:
      type: object
      properties:
        title:
          type: string
        genre:
          type: string
        thumbnailUrl:
          type: string
        primaryServerUrl:
          type: string
        backupServerUrl:
          type: string

    SiteSettings:
      type: object
      properties:
        websiteTitle:
          type: string
        motto:
          type: string
        specialFolderThumbnail:
          type: string
        specialFolderLabel:
          type: string
        countdownTargetDate:
          type: string
        facebook:
          type: string
        youtube:
          type: string
        telegram:
          type: string
        instagram:
          type: string
        dailymotion:
          type: string
        rumble:
          type: string
        whatsapp:
          type: string
        telegramChannel:
          type: string

    SettingsUpdate:
      type: object
      properties:
        websiteTitle:
          type: string
        motto:
          type: string
        specialFolderThumbnail:
          type: string
        specialFolderLabel:
          type: string
        countdownTargetDate:
          type: string
        facebook:
          type: string
        youtube:
          type: string
        telegram:
          type: string
        instagram:
          type: string
        dailymotion:
          type: string
        rumble:
          type: string
        whatsapp:
          type: string
        telegramChannel:
          type: string

    VoiceArtist:
      type: object
      required: [id, name, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        displayOrder:
          type: integer

    VoiceArtistInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        displayOrder:
          type: integer

    LoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1

    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string

    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 6