# lib/api-zod/src/generated/types/episodeStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TVR Dubbers streaming API
 * OpenAPI spec version: 0.1.0
 */
import type { Episode } from './episode';
import type { EpisodeStatsGenreCountsItem } from './episodeStatsGenreCountsItem';

export interface EpisodeStats {
  totalEpisodes: number;
  totalSpecial: number;
  totalViews: number;
  genreCounts: EpisodeStatsGenreCountsItem[];
  recentEpisodes: Episode[];
}