# lib/api-zod/src/generated/types/passwordChangeInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TVR Dubbers streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordChangeInput {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}