# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TVR Dubbers streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  websiteTitle?: string;
  motto?: string;
  specialFolderThumbnail?: string;
  specialFolderLabel?: string;
  countdownTargetDate?: string;
  facebook?: string;
  youtube?: string;
  telegram?: string;
  instagram?: string;
  dailymotion?: string;
  rumble?: string;
  whatsapp?: string;
  telegramChannel?: string;
}